+++ .tool-versions
@@ -2,4 +2,4 @@ ruby 2.6.5
 nodejs 12.16.1
 java zulu-8.56.0.21
 elixir 1.10.4
-erlang 23.0.3
+erlang 23.3.1
