+++ figx/test/figx_web/controllers/page_controller_test.exs
@@ -1,5 +1,5 @@
 defmodule FigxWeb.PageControllerTest do
-  use FigxWeb.ConnCase
+  use FigxWeb.ConnCase, async: true
 
   test "GET /", %{conn: conn} do
     conn = get(conn, "/")
